Refaktoroi: Siirrä metodit uudelleenkäytettäviksi
